🚧 Largely finished borders
