+++ src/PrettyTables.jl
@@ -80,7 +80,7 @@ include("./backends/text/text_backend.jl")
 
 include("./backends/excel/types.jl")
 #include("./backends/excel/documentation.jl")
-include("./backends/excel/helpers.jl")
+#include("./backends/excel/helpers.jl")
 include("./backends/excel/excel_backend.jl")
 
 include("./printing_state/alignment.jl")
